+++ pkg/R/Cwt_Thierry.R
@@ -56,8 +56,8 @@ cwtTh <- function(input, noctave, nvoice = 1, moments,
    dim(input) <- c(newsize,1)
 
    z <- .C("Scwt_thierry",
-           as.single(Re(input)),
-           as.single(Im(input)),
+           as.double(Re(input)),
+           as.double(Im(input)),
            Rtmp = as.double(Routput),
            Itmp = as.double(Ioutput),
            as.integer(noctave),
@@ -124,11 +124,11 @@ vwtTh <- function(input, scale, moments)
    dim(input) <- c(newsize,1)
 
    z <- .C("Svwt_Thierry",
-           as.single(Re(input)),
-           as.single(Im(input)),
+           as.double(Re(input)),
+           as.double(Im(input)),
            Rtmp = as.double(Routput),
            Itmp = as.double(Ioutput),
-           as.single(scale),
+           as.double(scale),
            as.integer(newsize),
            as.integer(moments),
            PACKAGE="Rwave")
